refactor(shared): decompose shared type contracts
