+++ packages/shared/src/types/repository-catalog.ts
@@ -1,0 +1,65 @@
+import type { ConfidenceLevel } from "./statuses";
+
+// Repository types for GitHub App installation
+export interface InstallationRepository {
+  id: number;
+  owner: string;
+  name: string;
+  fullName: string;
+  description: string | null;
+  private: boolean;
+  defaultBranch: string;
+  archived: boolean;
+  language?: string | null;
+  topics?: string[];
+}
+
+export interface RepoMetadata {
+  description?: string;
+  aliases?: string[];
+  channelAssociations?: string[];
+  keywords?: string[];
+  /**
+   * Environment opened by GitHub-bot sessions triggered from this repo
+   * (design §13.2). The bot falls back to a repo-bound session when the
+   * environment no longer exists or no longer contains this repository.
+   */
+  defaultEnvironmentId?: string;
+}
+
+export interface EnrichedRepository extends InstallationRepository {
+  metadata?: RepoMetadata;
+}
+
+// Bot package shared types
+export interface RepoConfig {
+  id: string;
+  owner: string;
+  name: string;
+  fullName: string;
+  displayName: string;
+  description: string;
+  defaultBranch: string;
+  private: boolean;
+  language?: string | null;
+  topics?: string[];
+  aliases?: string[];
+  keywords?: string[];
+  channelAssociations?: string[];
+}
+
+export type ControlPlaneRepo = EnrichedRepository;
+
+export interface ControlPlaneReposResponse {
+  repos: ControlPlaneRepo[];
+  cached: boolean;
+  cachedAt: string;
+}
+
+export interface ClassificationResult {
+  repo: RepoConfig | null;
+  confidence: ConfidenceLevel;
+  reasoning: string;
+  alternatives?: RepoConfig[];
+  needsClarification: boolean;
+}
